[feat] Add 'Challenge' Logic

+++ teamplan/Sources/Object/Project/ProjectObject.swift
@@ -13,7 +13,7 @@ import Foundation
 //============================
 struct ProjectObject{
     // id
-    let proj_id: Int64
+    let proj_id: Int
     
     // content
     var proj_title: String
@@ -36,7 +36,7 @@ struct ProjectObject{
     // Constructor
     // Coredata
     init(projectEntity: ProjectEntity){
-        self.proj_id = projectEntity.proj_id
+        self.proj_id = Int(projectEntity.proj_id)
         self.proj_title = projectEntity.proj_title ?? ""
         self.proj_started_at = projectEntity.proj_started_at ?? Date()
         self.proj_deadline = projectEntity.proj_deadline ?? Date()
@@ -53,7 +53,7 @@ struct ProjectObject{
     
     // DummyTest
     init(proj_id: Int64, proj_title: String, proj_started_at: Date, proj_deadline: Date, proj_finished: Bool, proj_todo: [TodoObject], proj_todo_registed: Int, proj_todo_finished: Int, proj_registed_at: Date, proj_changed_at: Date, proj_finished_at: Date) {
-        self.proj_id = proj_id
+        self.proj_id = Int(proj_id)
         self.proj_title = proj_title
         self.proj_started_at = proj_started_at
         self.proj_deadline = proj_deadline
@@ -72,7 +72,7 @@ struct ProjectObject{
 //============================
 struct TodoObject{
     // id
-    let todo_id: Int64
+    let todo_id: Int
     
     // content
     var todo_desc: String
@@ -89,7 +89,7 @@ struct TodoObject{
     // Costructor
     // Coredata
     init(todoEntity: TodoEntity){
-        self.todo_id = todoEntity.todo_id
+        self.todo_id = Int(todoEntity.todo_id)
         self.todo_desc = todoEntity.todo_desc ?? ""
         self.todo_pinned = todoEntity.todo_pinned
         self.todo_status = todoEntity.todo_status
@@ -100,7 +100,7 @@ struct TodoObject{
     
     // DummyTest
     init(todo_id: Int64, todo_desc: String, todo_pinned: Bool, todo_status: Bool, todo_registed_at: Date, todo_changed_at: Date, todo_updated_at: Date) {
-        self.todo_id = todo_id
+        self.todo_id = Int(todo_id)
         self.todo_desc = todo_desc
         self.todo_pinned = todo_pinned
         self.todo_status = todo_status
